+++ backend/expenses/Touir.ExpensesManager.Expenses/Controllers/DTO/FamilyNameDto.cs
@@ -1,0 +1,8 @@
+namespace Touir.ExpensesManager.Expenses.Controllers.DTO
+{
+    public class FamilyNameDto
+    {
+        public int Id { get; set; }
+        public string Name { get; set; } = null!;
+    }
+}